+++ docs/examples/database/adb/autonomous_database_wallet.tf
@@ -1,6 +1,6 @@
 // Copyright (c) 2017, 2019, Oracle and/or its affiliates. All rights reserved.
 
-resource "random_string" "wallet_password" {
+resource "random_string" "autonomous_database_wallet_password" {
   length  = 16
   special = true
 }
@@ -8,7 +8,7 @@ resource "random_string" "wallet_password" {
 data "oci_database_autonomous_database_wallet" "autonomous_database_wallet" {
   #Required
   autonomous_database_id = "${oci_database_autonomous_database.autonomous_database.id}"
-  password               = "${random_string.wallet_password.result}"
+  password               = "${random_string.autonomous_database_wallet_password.result}"
 }
 
 resource "local_file" "autonomous_database_wallet_file" {
@@ -16,6 +16,6 @@ resource "local_file" "autonomous_database_wallet_file" {
   filename = "${path.module}/autonomous_database_wallet.zip"
 }
 
-output "wallet_password" {
-  value = ["${random_string.wallet_password.result}"]
+output "autonomous_database_wallet_password" {
+  value = "${random_string.autonomous_database_wallet_password.result}"
 }