+++ docs/examples/database/adb/variables.tf
@@ -16,7 +16,19 @@ variable "autonomous_database_data_storage_size_in_tbs" {
 }
 
 variable "autonomous_database_db_name" {
-  default = "atpdb1"
+  default = "adbdb1"
+}
+
+variable "autonomous_data_warehouse_db_name" {
+  default = "adbdw1"
+}
+
+variable "autonomous_database_db_workload" {
+  default = "OLTP"
+}
+
+variable "autonomous_data_warehouse_db_workload" {
+  default = "DW"
 }
 
 variable "autonomous_database_defined_tags_value" {
@@ -25,6 +37,10 @@ variable "autonomous_database_defined_tags_value" {
 
 variable "autonomous_database_display_name" {
   default = "example_autonomous_database"
+}
+
+variable "autonomous_data_warehouse_display_name" {
+  default = "example_autonomous_data_warehouse"
 }
 
 variable "autonomous_database_freeform_tags" {